+++ src/sync_connection_wrapper/mod.rs
@@ -221,7 +221,6 @@ mod implementation {
         type TransactionManager =
             SyncTransactionManagerWrapper<<C as Connection>::TransactionManager>;
 
-        #[cfg(not(target_arch = "wasm32"))]
         async fn establish(database_url: &str) -> ConnectionResult<Self> {
             let database_url = database_url.to_string();
             let mut runtime = S::get_runtime();